AG-17082 Fix "lint:circular"
