+++ packages/ag-charts-community/src/chart/series/series.ts
@@ -70,6 +70,7 @@ import { Observable } from '../../util/observable';
 import type { ChartAxis } from '../chartAxis';
 import type { ChartMode } from '../chartMode';
 import type { DataController } from '../data/dataController';
+import type { DataModel, ProcessedData } from '../data/dataModel';
 import { DataSet } from '../data/dataSet';
 import type { ChartLegendDatum, ChartLegendType } from '../legend/legendDatum';
 import type { Marker } from '../marker/marker';
@@ -84,7 +85,6 @@ import type {
     DatumIndexType,
     INodeEvent,
     ISeries,
-    ISeriesEventsMaps,
     ISeriesProperties,
     NodeDataDependencies,
     SeriesNodeDatum,
@@ -92,6 +92,11 @@ import type {
 } from './seriesTypes';
 import { type ShapeFillBBox } from './shapeUtil';
 import { hasDimmedOpacity, resolveMarkerDrawingMode } from './util';
+
+export interface SeriesDataEvent {
+    readonly dataModel: DataModel<any, any, any>;
+    readonly processedData: ProcessedData<any>;
+}
 
 /** Modes of matching user interactions to rendered nodes (e.g. hover or click) */
 export enum SeriesNodePickMode {
@@ -531,7 +536,11 @@ export abstract class Series<
         };
     }
 
-    readonly events = new EventEmitter<ISeriesEventsMaps>();
+    readonly events = new EventEmitter<{
+        'data-update': SeriesDataEvent;
+        'data-processed': SeriesDataEvent;
+        'data-selection-change': null;
+    }>();
 
     override addEventListener(type: 'seriesVisibilityChange', listener: (e: AgSeriesVisibilityChange) => void): void;
     override addEventListener(type: 'seriesNodeClick', listener: (e: SeriesNodeEvent<any>) => void): void;
